\name{cod_bases}
\alias{cod_bases}
\title{Diccionario de Bases de datos Pulso Social
}
\description{
Un conjunto de datos que contiene información sobre múltiples bases de datos de Social Pulso.}
\usage{
cod_bases
}
\details{
   \bold{id_data:} {Un identificador único para cada base de datos. Se trata de un campo numérico.}

   \bold{db_nombre:} {Nombre de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_descripcion:} {Descripción de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_descarga:} {Enlace de descarga de la base de datos. Se trata de un campo de caracteres.}

   \bold{Actualizacion:} {Fecha de la última actualización de la base de datos. Se trata de un campo Fecha.}

   \bold{db_fuente:} {La fuente de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_link:} {Un enlace a la base de datos. Se trata de un campo de caracteres.}

   \bold{db_nivel:} {El nivel de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_territorial:} {Ámbito territorial de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_poblacional:} {El ámbito de población de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_poblacional2:} {Ámbito de población adicional de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_periocidad:} {Frecuencia de actualización de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_disponibilidad:} {La disponibilidad de la base de datos. Se trata de un campo de caracteres.}

   \bold{db_disponibilidad2:} {Información adicional sobre la disponibilidad de la base de datos. Se trata de un campo de caracteres.}

}
\author{
}
\seealso{
}

