\name{cod_vars}
\alias{cod_vars}
\title{Diccionario de Variables Sociales Pulso
}
\description{
Un conjunto de datos que contiene información sobre múltiples variables de Pulso Social.
}
\usage{
cod_vars
}
\details{
   \bold{var_id:} {Un identificador único para cada variable. Se trata de un campo numérico.}

   \bold{id_data:} {Un identificador único para cada base de datos. Es un campo numérico.}

   \bold{db_nombre:} {Nombre de la base de datos. Se trata de un campo de caracteres.}

   \bold{variable:} {Nombre de la variable. Es un campo de caracteres.}

   \bold{var_etiqueta:} {La etiqueta de la variable. Es un campo de caracteres.}

   \bold{var_descripcion:} {Descripción de la variable. Es un campo de caracteres.}

   \bold{var_unidad:} {Unidad de medida de la variable. Es un campo de caracteres.}

   \bold{var_periodos:} {Los periodos de la variable. Es un campo de caracteres.}

   \bold{var_niveles:} {Los niveles de la variable. Este es un campo de caracteres.}

   \bold{var_terri:} {El ámbito territorial de la variable. Es un campo de carácter.}

   \bold{var_poblacional:} {Ámbito poblacional de la variable. Es un campo de carácter.}

   \bold{Tipo:} {El tipo de la variable. Es un campo de caracteres.}

   \bold{dimension:} {La dimensión de la variable. Es un campo de caracteres.}

   \bold{subdimension:} {La subdimensión de la variable. Es un campo de caracteres.}
}
\author{
}
\seealso{
}



