\name{map_dptos}
\alias{map_dptos}
\title{Mapas de deparamentos de Colombia
}
\description{
Este Dataset lee un shapefile que contiene datos cartográficos de los departamentos colombianos, con los nombres de convención utilizados en el resto del proyecto Pulso Social.
}
\usage{
map_dptos
}
\format{
}

\author{
}
\seealso{

}

