\name{map_mpios}
\alias{map_mpios}
\title{Colombia Municipalities Maps Data

}
\description{
This Dataset reads a shapefile containing map data of Colombian Municipalities, with convention names used in the rest of the Pulso Social project.
}
\usage{
map_mpios
}
\format{
}

\author{
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

