\name{nom_dptos}
\alias{nom_dptos}
\title{Nombres de deparamentos en Colombia
}
\description{
Este conjunto de datos lee un archivo que contiene datos de nombres de departamentos colombianos, con nombres de convención utilizados en el resto del proyecto Pulso Social.
}
\usage{
nom_dptos
}
\format{
}

\author{
}
\seealso{
}



