\name{nom_mpios}
\alias{nom_mpios}
\title{Colombia Municipalities Name Data
}
\description{
This Dataset reads a file containing data on the names of Colombian Municipalities, with convention names used in the rest of the Pulso Social project.

}
\usage{
nom_mpios
}
\format{
}

\author{
}
\seealso{
}
