\name{nom_regiones}
\alias{nom_regiones}
\title{Colombia Regions Name Data
}
\description{
This Dataset reads a file containing data on the names of Colombian Regions, with convention names used in the rest of the Pulso Social project.

}
\usage{
nom_regiones
}
\format{
}

\author{
}
\seealso{
}
