% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_pulso_dic.R
\name{pulso_diccionario}
\alias{pulso_diccionario}
\title{Dictionary of Pulso Social Colombia}
\usage{
pulso_diccionario()
}
\value{
Updates the dictionary for Pulso Social Colombia
}
\description{
Dictionary of Pulso Social Colombia
}
\examples{
pulso_diccionario()
}
