% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pulso_map.R
\name{pulso_map}
\alias{pulso_map}
\title{Comparative Map Analysis for Pulso Social}
\usage{
pulso_map(id, type_p, col_palette)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of graph (optional, default is "").}

\item{col_palette}{Custom color palette (optional).}
}
\value{
A ggplot object that displays the map graph comparing the maximum and minimum years of the "id" variable.
}
\description{
This function generates a comparative map graph between the maximum and minimum years for a specific variable in the data of Pulso Social.
}
\examples{
pulso_map(id="286", type_p="print")
pulso_map(id="300", type_p="print", col_palette=c("#7052ff", "#d06aff", "#72A3AF", "#ff81ff", "#5B6477", "#ff81ff"))
}
