% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_pulso_scatter.R
\name{pulso_scatter}
\alias{pulso_scatter}
\title{Scatter Analysis of Pulso Social}
\usage{
pulso_scatter(id, type_p, year, col_palette)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of graph (optional, default is "").}

\item{year}{Year for which you want to generate the graph (optional, by default, the last available year is selected).}

\item{col_palette}{Custom color palette (optional).}
}
\value{
A ggplot object that shows the scatter plot of the "id" variable.
}
\description{
This function generates a scatter plot that displays the relationship between two categories of a specific variable in the data of Pulso social for a specific year.
}
\examples{
pulso_scatter(id="46", type="print")
pulso_scatter(id="46", type="print", col_palette=c("#7052ff", "#d06aff", "#72A3AF", "#ff81ff", "#5B6477", "#ff81ff", "#60ff73", "#ccff60"))
}
