% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_pulso_scatter_time.R
\name{pulso_scatter_time}
\alias{pulso_scatter_time}
\title{Time Series Scatter Analysis for Pulso Social}
\usage{
pulso_scatter_time(id, type_p, year_min, year_max, col_palette)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of plot (optional, default is "").}

\item{year_min}{Minimum year for the year range of the plot, defaults to the minimum year in the data.}

\item{year_max}{Maximum year for the year range of the plot, defaults to the maximum year in the data.}

\item{col_palette}{Custom color palette (optional).}
}
\value{
A ggplot object that displays the scatter plot between two years of the variable "id."
}
\description{
This function generates a scatter plot that displays the relationship between two years of a specific variable "id" in the Pulso Social data.
}
\examples{
pulso_scatter_time(id="300",type="print")
pulso_scatter_time(id="300",type="print",2010,2020)
pulso_scatter_time(id="300",type="print",2010,2020,c("#7052ff", "#d06aff", "#72A3AF", "#ff81ff", "#5B6477", "#ff81ff", "#60ff73", "#ccff60"))
}
