% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pulso_static.R
\name{pulso_static}
\alias{pulso_static}
\title{Static Analysis of Pulso Social}
\usage{
pulso_static(id, type_p, year)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of graph (optional, default is "").}

\item{year}{Year for which to generate the analysis (optional, default is the most complete year of data).}
}
\value{
A ggplot object that displays the static graph of the "id" variable at a certain time.
}
\description{
This function generates a static circular bar chart for a specific variable in the data of Pulso So.
}
\examples{
pulso_static(id="286",type="print")
pulso_static(id="290",type="",2020)
}
