% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pulso_trend.R
\name{pulso_trend}
\alias{pulso_trend}
\title{Pulso Social Colombia - Gráfico de tendencias}
\usage{
pulso_trend(id, type_p)
}
\arguments{
\item{id}{Identificación de la variable de Pulso Social Colombia}

\item{type_p (opcional)}{Provee la opción de agregar las etiquetas y variables usando la opción "print". Este parámetro está predeterminado para estar vacio, lo cual implica que no se imprime las etiquetas.}
}
\value{
Objeto ggplot que muestra el gráfico de tendencia de la variable "id" con datos de Pulso Social - Colombia. 
}
\description{
Esta función crear un gráfico de tendencias con los variables existentes en la infraestructura de pulso social. En el caso que la variable no tenga información temporal un mensaje de advertencia 
}
\examples{
pulso_trend(id="297",type="print") # imprimen todo los títulos de la imagen
pulso_trend(id="290",type="") # imprime la gráfica sin títulos o etiquetes, esta es la función predeterminada
pulso_trend(106) # imprime la gráfica sin títulos o etiquetes
}
