\name{cod_bases}
\alias{cod_bases}
\title{Dictionary of Pulso Social Datasets
}
\description{
A dataset containing information about multiple Pulso Social Project databases.
}
\usage{
cod_bases
}
\details{
   \bold{id_data:} {A unique identifier for each database. This is a numeric field.}

   \bold{db_nombre:} {The name of the database. This is a character field.}

   \bold{db_descripcion:} {A description of the database. This is a character field.}

   \bold{db_descarga:} {A download link for the database. This is a character field.}

   \bold{Actualizacion:} {The date of the last update of the database. This is a Date field.}

   \bold{db_fuente:} {The source of the database. This is a character field.}

   \bold{db_link:} {A link to the database. This is a character field.}

   \bold{db_nivel:} {The level of the database. This is a character field.}

   \bold{db_territorial:} {The territorial scope of the database. This is a character field.}

   \bold{db_poblacional:} {The population scope of the database. This is a character field.}

   \bold{db_poblacional2:} {Additional population scope of the database. This is a character field.}

   \bold{db_periocidad:} {The frequency of database updates. This is a character field.}

   \bold{db_disponibilidad:} {The availability of the database. This is a character field.}

   \bold{db_disponibilidad2:} {Additional availability information of the database. This is a character field.}


}
\author{
}
\seealso{
}

