\name{cod_vars}
\alias{cod_vars}
\title{Dictionary of Pulso Social Variables
}
\description{
A dataset containing information about multiple Pulso Social Project variables.
}
\usage{
cod_vars
}
\details{
   \bold{var_id:} {A unique identifier for each variable. This is a numeric field.}

   \bold{id_data:} {A unique identifier for each database. This is a numeric field.}

   \bold{db_nombre:} {The name of the database. This is a character field.}

   \bold{variable:} {The name of the variable. This is a character field.}

   \bold{var_etiqueta:} {The label of the variable. This is a character field.}

   \bold{var_descripcion:} {A description of the variable. This is a character field.}

   \bold{var_unidad:} {The unit of measurement for the variable. This is a character field.}

   \bold{var_periodos:} {The periods of the variable. This is a character field.}

   \bold{var_niveles:} {The levels of the variable. This is a character field.}

   \bold{var_terri:} {The territorial scope of the variable. This is a character field.}

   \bold{var_poblacional:} {The population scope of the variable. This is a character field.}

   \bold{Tipo:} {The type of the variable. This is a character field.}

   \bold{dimension:} {The dimension of the variable. This is a character field.}

   \bold{subdimension:} {The subdimension of the variable. This is a character field.}
}
\author{
}
\seealso{
}



