\name{map_dptos}
\alias{map_dptos}
\title{Colombia Departments Maps Data
}
\description{
This Dataset reads a shapefile containing map data of Colombian departments, with convention names used in the rest of the Pulso Social project.
}
\usage{
map_dptos
}
\format{
}

\author{
}
\seealso{

}

