\name{map_mpios}
\alias{map_mpios}
\title{Mapas de Municipios de Colombia

}
\description{
Este Dataset lee un shapefile que contiene datos de mapas de Municipios Colombianos, con nombres de convención usados en el resto del proyecto Pulso Social.
}
\usage{
map_mpios
}
\format{
}

\author{
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

