\name{nom_dptos}
\alias{nom_dptos}
\title{Colombia Departments Name Data
}
\description{
This Dataset reads a file containing Colombian department name data, with convention names used in the rest of the Pulso Social project.
}
\usage{
nom_dptos
}
\format{
}

\author{
}
\seealso{
}



