\name{nom_mpios}
\alias{nom_mpios}
\title{Nombre de municipios en Colombia
}
\description{
Este Dataset lee un archivo que contiene datos sobre los nombres de los Municipios Colombianos, con nombres de convención utilizados en el resto del proyecto Pulso Social.

}
\usage{
nom_mpios
}
\format{
}

\author{
}
\seealso{
}
