\name{nom_regiones}
\alias{nom_regiones}
\title{Nombres Regiones en Colombia
}
\description{
Este conjunto de datos lee un archivo que contiene datos sobre los nombres de las regiones colombianas, con los nombres de convención utilizados en el resto del proyecto Pulso Social.

}
\usage{
nom_regiones
}
\format{
}

\author{
}
\seealso{
}
