% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pulso_map_change.R
\name{pulso_map_change}
\alias{pulso_map_change}
\title{Map Analysis for Percentage Change in Pulso Social}
\usage{
pulso_map_change(id, type_p, col_palette)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of plot (optional, default is "").}

\item{col_palette}{Custom color palette (optional).}
}
\value{
A ggplot object that displays the map plot of percentage change between the maximum and minimum years.
}
\description{
This function generates a map plot that displays the percentage change of a specific variable in the Pulso Social data between the maximum and minimum years.
}
\examples{
pulso_map_change(id="286",type="print")
pulso_map_change(id="300",type="print",c("#7052ff", "#d06aff", "#72A3AF", "#ff81ff", "#5B6477", "#ff81ff"))
}
