% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_pulso_scatter.R
\name{pulso_scatter}
\alias{pulso_scatter}
\title{Análisis de dispersión por niveles}
\usage{
pulso_scatter(id, type_p, year, col_palette)
}
\arguments{

\item{id}{Identificación de la variable de Pulso Social Colombia}

\item{type_p (opcional)}{Provee la opción de agregar las etiquetas y variables usando la opción "print". Este parámetro está predeterminado para estar vacio, lo cual implica que no se imprime las etiquetas.}

\item{year (opcional)}{Año que se desea realizar el gráfico Year for which to generate the analysis. El caso de no proveerse se produce el gráfico con los últimos datos disponibles.}

\item{col_palette (optional)}{Permite la definición pre-determinada de los colores que se utilizarán para el mapa. El insumo debe definir \bold{5} colores. }

}
\value{
    Objeto ggplot que muestra el gráfico de dispersión de un "id" con datos de Pulso Social - Colombia. 
}
\description{
    Esta función genera un gráfico de dispersión por niveles gráfico de dispersión que muestra la relación entre dos categorías de una variable específica en los datos de Pulso social para un año concreto.
}
\examples{
pulso_scatter(id="46", type="print") # Imprime las etiquetas y notas
pulso_scatter(id="46", type="print", col_palette=c("#7052ff", "#d06aff", "#72A3AF", "#ff81ff", "#5B6477", "#ff81ff", "#60ff73", "#ccff60")) # Imprime las etiquetas y notas
pulso_scatter(46)
}
