% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_pulso_trend.R
\name{pulso_trend}
\alias{pulso_trend}
\title{Function for Pulso Social Trend Analysis}
\usage{
pulso_trend(id, type_p)
}
\arguments{
\item{id}{Unique identifier of the variable.}

\item{type_p}{Type of plot: "print" includes titles and notes about the data. Default is an empty string, which displays the plot without titles or notes.}
}
\value{
A ggplot object that displays the trend graph of the "id" variable with data from Pulso Social - Colombia.
}
\description{
Function for Pulso Social Trend Analysis
}
\examples{
pulso_trend(id="297",type="print")
pulso_trend(id="290",type="")
pulso_trend(106)
}
